Remove CMake workflow and renamed craft_initialize_key function
This commit does two major changes. First, it removes the entire CMake configuration workflow (cmake.yml) found in the .github/workflows directory since it's no longer needed in our build workflow.

Second, it renames the function "craft_initialize_key" to "craft_key_schedule" in the files craft.c and craft.h found in the src/craft directory. The renaming enhances code readability, as it provides a more accurate description of the function's role in scheduling the key in the Craft lightweight cryptographic algorithm.